Got a good design going
I have a bad habit of forgetting that tailwind is now tailwindcss so that was rather embarrassing on my part. Regardless it is fixed now, and the core app design looks quite nice. Still need to add more stuff but for the time being this is a good start that I'd consider ready for preview.