Working on Authentication for Users atm
This is probably one of my least favourite parts of development unfortunately, but I will try to keep at it. For now, I had the idea to make it so that Quotes are able to show a list of the users who liked them, (users are able to make their likes private, and if they choose to do so they will not be displayed here). A simple idea, but a good one :)