Tests updated
Had to change the entire approach since the auths were not playing nicely. Using mjs instead of ts as it made it easier and this was already irritating me. Lazy sure, but there's no reason that they can't be fixed in future. I just really did not like doing this part pretty much.