Added ability to change name and email, and make likes private
Currently a sort of glitch going on with browser autofill when I was testing some of the features and trying to use an already-in-use account, but I think that might be more the browser/autofill's fault than anything, as it worked perfectly when manually entered. Will try getting some more features released before Christmas.