Added searching for quotes (from the main screen)
I promised I'd deliver before Christmas! I Added a search feature that shows the first 50 results for quotes searched that match, author, category, or tags (I thought about just a general regex, but i think that would broaden it a bit too much in terms of what we want. I'll change in future if this proves unsatisfactory). Had a weird issue with the Prisma client that I needed to patch up too, but that's just a minor thing.

Merry Christmas everyone! :)